% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_typhoon_plot_functions.R
\name{DensDotPlot}
\alias{DensDotPlot}
\title{Makes a heatmap of data within a matrix}
\usage{
DensDotPlot(dot.m, z_min = 0, z_max = 100, low_col = "white",
  med_col = "", high_col = "blue", num_colors = 100, plot_title = "",
  plot_title_line = NA, x_label = "", y_label = "", x_axt = "s",
  y_axt = "s", plot_box = TRUE)
}
\arguments{
\item{dot.m}{matrix, the 2-dimensional matrix to plot}

\item{z_min}{numeric, the minimum intensity of the plot (default: 0)}

\item{z_max}{numeric, the maximum intensity of the plot (default: 50)}

\item{low_col}{character, the minimum color (default: "white")}

\item{med_col}{character, the middle color (optional, default: "")}

\item{high_col}{character, the maximum color (default: "blue")}

\item{num_colors}{numeric, the number of colors between low_col and high_col}

\item{plot_title,x_label,y_label}{character (optional, default: "" for all)}

\item{plot_title_line}{numeric, (optional, default: NA)}

\item{x_axt,y_axt}{character, "s" (default) or "n"}

\item{plot_box}{boolean, (default: TRUE)}
}
\value{
A heatmap plot utilizing the image function
}
\description{
Makes a heatmap of data within a matrix
}
\examples{
\dontrun{DensDotPlot(dot.m = typhoon.m, z_max = 50)}
}

