% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_typhoon_plot_functions.R
\name{GetAggregateVPlot}
\alias{GetAggregateVPlot}
\title{Obtain the aggregate VPlot matrix over a set of chromosomal coordinates (e.g. TF-binding sites)}
\usage{
GetAggregateVPlot(data.df, bam_file_name, win)
}
\arguments{
\item{data.df}{Data Frame, a feature data frame containing the "chr", "mid", and "strand" components for each location in the dataset}

\item{bam_file_name}{Character, a bam file and its associated .bai index file}

\item{win}{Numeric, number of bp to extend the V-plot from the midpoint location}
}
\value{
A frag_length x position matrix (250 rows x (2*win + 1) columns) that
   can be visualized with DensDotPlot
}
\description{
Given a data frame listing chromosomal coordinates, an indexed bam file, 
and a bp window, this function will return a matrix displaying the paired-end, 
sequenced reads in a standard V-plot (e.g. at the midpoint-fragment length
location in a matrix.
}
\examples{
\dontrun{GetAggregateVPlot(abf1_yeast_sacCer2.df, "/home/jab112/dm242.bam", 500)}
}
