% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_typhoon_plot_functions.R
\name{GetTyphoonPlotMat}
\alias{GetTyphoonPlotMat}
\title{Get the TyphoonPlot matrix}
\usage{
GetTyphoonPlotMat(bam_file_name, chr, start_pos, end_pos)
}
\arguments{
\item{bam_file_name}{character, a bam file and its associated .bai index file}

\item{chr}{character, chromosome from the bam file}

\item{start_pos}{numeric, genomic start position}

\item{end_pos}{numeric, genomic end position}
}
\value{
A frag_length x position matrix (250 rows x position columns) that
 can be visualized with DensDotPlot
}
\description{
Get the TyphoonPlot matrix
}
\examples{
\dontrun{GetTyphoonPlotMat("/home/jab112/dm242.bam", "1", 10000, 15000)}
}

