% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_typhoon_plot_functions.R
\name{PlotGene}
\alias{PlotGene}
\title{Helper function to plot individual genes as rectangles.}
\usage{
PlotGene(gene.v, y_low, y_high, x_start, x_end, cex_title, geneName,
  x_pos_title = 50, fwd_gene_color, rev_gene_color)
}
\arguments{
\item{gene.v}{Vector, selected row from \code{yeast_gene.df} feature file
containing information about the gene.}

\item{y_low}{Numeric, the bottom \emph{y} rectangular coordinate.}

\item{y_high}{Numeric, the top \emph{y} rectangular coordinate.}

\item{x_start}{Numeric, the left \emph{x} rectangular coordinate of the plot.}

\item{x_end}{Numeric, the right \emph{x} rectangular coordinate of the plot.}

\item{cex_title}{Numeric, specifies the cex expansion factor for the title}

\item{geneName}{Boolean, should the gene name be displayed (Default: T)}

\item{x_pos_title}{Numeric, the number of inset bp specifying the gene name.}
}
\value{
Rectangle schematic depicting gene coordinates
}
\description{
Helper function for \code{\link{MakeGeneSchematic}} to plot individual
genes.  Each gene is designated as a gray box, with genes on the Watson (+)
strand displayed in the top row and genes on the Crick (-) strand displayed
on the bottom row.  Orientation of the gene name also denotes transcription
direction.
}
\examples{
 PlotGene(gene.v = yeast_gene.df[1,], y_low = 0.5, y_high = 1,
          x_start = 40000, x_end = 41000, cex_title = 1,
          geneName = T, x_pos_title = 50,
          fwd_gene_color = "gray", rev_gene_color = "gray"
         )
}
