% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TyphoonPlot.R
\docType{package}
\name{TyphoonPlot}
\alias{TyphoonPlot}
\alias{TyphoonPlot-package}
\title{TyphoonPlot: A package for creating a TyphoonPlot}
\description{
The TyphoonPlot package provides helper processing and graphics functions to
display paired-end next-generation sequencing data in a helpful visualization
format
}
\section{Required Packages}{

This package requires the following packages:
 gplots
 GenomicRanges
 Rsamtools
}

\section{TyphoonPlot Processing Functions}{

 GetTyphoonPlotMat
 GetTyphoonPlotMatJava
}

\section{TyphoonPlot Graphing Functions}{

 DensDotPlot
}

