% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_typhoon_plot_functions.R
\name{DensDotPlot}
\alias{DensDotPlot}
\title{Makes a heatmap of data within a matrix}
\usage{
DensDotPlot(dot.m, z_min = 0, z_max = NA, low_col = "white",
  med_col = "", high_col = "blue", num_colors = 100, plot_title = "",
  plot_title_line = NA, x_label = "", y_label = "", x_axt = "s",
  y_axt = "s", plot_box = TRUE)
}
\arguments{
\item{dot.m}{Matrix, the 2-dimensional matrix to plot}

\item{z_min}{Numeric, the minimum intensity of the plot (default: 0)}

\item{z_max}{Numeric, the maximum intensity of the plot.  If not specified,
by default the maximum will be set to the value at the 95th percentile (default: NA)}

\item{low_col}{Character, the minimum color (default: "white")}

\item{med_col}{Character, the middle color (optional, default: "")}

\item{high_col}{Character, the maximum color (default: "blue")}

\item{num_colors}{Numeric, the number of colors between low_col and high_col}

\item{plot_title, x_label, y_label}{Character (optional, default: "" for all)}

\item{plot_title_line}{Numeric, (optional, default: NA)}

\item{x_axt, y_axt}{Character, "s" (default) or "n"}

\item{plot_box}{Boolean, (default: TRUE)}
}
\value{
A heatmap plot utilizing the image function
}
\description{
This function takes as input a TyphoonPlot matrix created from
\code{\link{GetTyphoonPlotMat}} and uses the \code{\link[graphics]{image}}
function to create the heatmap.
}
\examples{
\dontrun{DensDotPlot(dot.m = typhoon.m, z_max = 50)}
}
