% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_typhoon_plot_functions.R
\name{GetTyphoonPlotMat}
\alias{GetTyphoonPlotMat}
\title{Obtain the TyphoonPlot matrix needed for plotting.}
\usage{
GetTyphoonPlotMat(bam_file_name, chr, start_pos, end_pos)
}
\arguments{
\item{bam_file_name}{Character, a bam file and its associated .bai index file}

\item{chr}{Character, chromosome from the bam file}

\item{start_pos}{Numeric, genomic start position}

\item{end_pos}{Numeric, genomic end position}
}
\value{
A frag_length x position matrix (250 rows x position columns) that
   can be visualized with DensDotPlot
}
\description{
Given an indexed bam file and chromosomal genomic coordinates, this function
will return a matrix displaying the paired-end, sequenced reads in
two dimensions.  Note that this function will plot reads at half their
span to reduce the overlap between adjacent reads.  For example, a 100 bp
paired-end read spanning from position 201 to 300 will fill a 50 bp
x-coordinate region from position 226 to 275.
}
\examples{
\dontrun{GetTyphoonPlotMat("/home/jab112/dm242.bam", "1", 10000, 15000)}
}

